Amounts size limitations removed.
Numeric instead int type of db filed used.

+++ src/patches/4.py
@@ -12,11 +12,11 @@ def apply(curs):
             currency_id int NOT NULL,
             FOREIGN KEY(currency_id) REFERENCES currency(id),
             creation_date timestamp with time zone NOT NULL DEFAULT now(),
-            available_real_amount int DEFAULT 0,
-            available_virtual_amount int DEFAULT 0,
+            available_real_amount NUMERIC DEFAULT 0,
+            available_virtual_amount NUMERIC DEFAULT 0,
             locking_order varchar[] DEFAULT NULL,
-            locked_amount int DEFAULT 0,
-            overdraft_limit int DEFAULT 0
+            locked_amount NUMERIC DEFAULT 0,
+            overdraft_limit NUMERIC DEFAULT 0
         )
     ''')
