+++ src/patches/6.py
@@ -11,8 +11,8 @@ def apply(curs): #IGNORE:W0622
             order_id varchar NOT NULL,
             order_type varchar,
             locking_date timestamp with time zone NOT NULL DEFAULT now(),
-            real_amount int,
-            virtual_amount int
+            real_amount NUMERIC,
+            virtual_amount NUMERIC
         )
     ''')
