+++ src/patches/8.py
@@ -12,8 +12,8 @@ def apply(curs):
             order_type varchar,
             locking_date timestamp with time zone NOT NULL,
             chargeoff_date timestamp with time zone NOT NULL DEFAULT now(),
-            real_amount int,
-            virtual_amount int
+            real_amount NUMERIC,
+            virtual_amount NUMERIC
         )
     ''')
